Add ability for editors to delete content from the edit modal

+++ client/src/pages/ContentPage/Card/ReviewCard.js
@@ -95,6 +95,10 @@ function ReviewCard(props) {
       return;
     }
 
+    if (!window.confirm("Confirm one last time that you want to delete the the proposed changes.")) {
+      return;
+    }
+
     // delete proposed changes
     const results = await fetch(`${API_URL}/cards/${props.card.cardId}/changes`, {
       method: "DELETE",
